Remove object references support
See the docs for this feature: https://w.wiki/$Z8
AFAIK this was only used for German Wiki Loves Monuments. The
feature is weirdly specific and hard to apply elsewhere, or even
explain. I think the Germans have given up on it as well, judging
from the more recent WLM instances.

Support for similar uploader customization could be provided in a
more general way in the future. That would need a real use case,
though.

Bug: T274910
Change-Id: I3ad3dff58796f072e730bca93e41cea7bd98109b