Add anchor page for global config
This anchor page (Campaign:-) doesn't contain any meaningful content,
it's just there so we have somewhere to attach template links used
in the global parsed config. By doing this, the anchor will get a
LinksUpdate every time its dependencies change and this in turn will
cause global parsed config cache invalidation. The mechanism also
works the other way around: when GlobalParsedConfig detects that the
config to be parsed has changed, GlobalConfigAnchorUpdateJob is
enqueued which will either create the anchor (if it doesn't exist) or
poke it with a null edit.

The code should not get stuck in an endless cache invalidation loop,
there are safeguards for this both in CampaignHooks and
GlobalParsedConfig.

Additionally:
* Refactored a good portion of the hooks to use DI. Most of them
  are now covered by unit tests.
* Introduced ConfigCacheInvalidator, which decouples cache
  invalidation functionality from the actual config classes, making
  it possible to invalidate config cache without loading the config
  and injecting some heavy services.
* The anchor page cannot be deleted or moved, but nothing will go bad
  if either happens for whatever reason.

Bug: T276425
Change-Id: I09622dc1f543c1596a357cbd9871cc9598050f0c