Localize the [[User:…]] link for own work uploads
This is a very minor thing, but previously these links were in
English only which was a bit confusing and inconsistent on wikis in
other languages. By first constructing a Title object and then
retrieving the prefixed title from there, we get a localized version
of the User: namespace.

This of course depends only on the wiki's language settings, not
user preferences.

Change-Id: I5c3a0a2656671a4c0ad84240c7d9e8c4acbd2c2e