Rewrite campaign config validation
Rewrote campaign config validation to work properly. The schema is
now defined in a more readable YAML file, complete with comments.
A small part of the schema (license list) is injected in PHP before
validation. The schema is cached in APCU.

MediaUploader will not allow saving campaigns with either invalid
markup or content that does not validate against the schema.

IMPORTANT: THIS SCHEMA BREAKS BACKWARDS-COMPATIBILITY WITH
UploadWizard. A follow-up commit will provide a maintenance script
for removing disallowed properties from old campaigns, with optional
conversion to YAML and code prettifying.

Bug: T277535

Change-Id: Ia9befb70fc3ed988a4489e40dd53df7579252f0a