Add maintenance script for fixing invalid campaigns
FixCampaigns.php is an update/utility script that should be useful
to anyone migrating from UploadWizard. It identifies superfluous
fields in campaigns and simply removes them, to make them pass
schema validation. It can also prettify all campaigns.

It might be possible to add more types of automatic fixes, but I
wasn't able to think of any at the moment. The implemented fix
seems the most pressing, as the new schema disallows additional
fields.

The script is hooked up to update.php to run once, but can safely
be re-ran as many times as you desire.

There might be some issues with how it behaves with certain
campaign definitions, but this is due to issues in the schema that
will be addressed later.

Similarly, issues related to YAML parsing and dumping should be
fixed upstream. They were reported to Symfony maintainers.

Bug: T277535
Change-Id: I7facd37805b845c796e66999f981b3c1224955f1