Add Title retrieval to CampaignSelectQueryBuilder
The builder is now able to retrieve records of campaigns along with
their corresponding titles. This will be very useful in several places
that require not only the ID of the campaign but also its title.

This is almost the last thing factored out of the old
UploadWizardCampaign class. The only thing left now are the stats,
which I'll do in the next patch.

The AllCampaigns API will stop returning entire campaign configs.
This is intentional. The ability to fetch campaign configs will be
moved to a different endpoint.

Bug: T276242
Change-Id: I6bce49047a572af6a82d665e9c1636f397890363