Introduce CampaignStats class
This should be the last part of refactoring the campaign subsystem.
The stats have now a simplified interface and are cached more
efficiently. They are also retrieved in batch, which should speed
things up for stuff like the allcampaigns API.

A lot of other refactors are included here as well.

Bug: T276242
Change-Id: I3155ce452df5f8511f91e9ebda62a8557695e72f