Rework field configuration in the details step
Made the field configuration thing consistent. Added the option to:
- enable/disable fields
- make field optional/recommended/required
- change the text labels
- change the available options in a field
- add a few new field types
- hide a field
- move a field to the "auxiliary" section at the bottom at the form
and more.

The relevant campaign settings are validated with a JSON Schema.

I hope it all works.

Currently broken is filling in the wikitext of the uploaded files.
This will be fixed separately as part of T275027.

Removed feature: warning about dates unlikely for public domain
content. The feature relied on Commons-specific license templates.

Bug: T275264
Bug: T274921
Change-Id: I9b4fa5789a4aeff13c1da7d6aa3fc3581e67bb69