Replace various deprecated methods
I can't really tell what the previously used getGlobalBlock() was
meant to do. The deprecation notice says it should be replaced with
getBlock(). But this made the SpecialMediaUploaderTest fail. The
test appears to make a difference between a "sitewide" block vs.
a "partial" block. But this is already properly respected by the
existing isBlockedFromUpload() call above. An additional getBlock()
call is apparently not needed.

Additionally, fix the test fixture of ConfigParserTest to update the
mock after the Mediawiki ParserOutput::getText replacement.

These changes should hopefully fix the extension CI.

Bug: T387574
Co-authored-by: Isabelle Hurbain-Palatin <ihurbainpalatin@wikimedia.org>
Change-Id: Ib763bcd570aec525347bf5febf3b9ec0ac96c83f